+++ src/layout/components/Navbar/index.vue
@@ -1,26 +1,44 @@
 <template>
   <div class="navbar">
-    <hamburger :is-active="ctx.sidebar.opened" @toggleClick="toggleSidebar" class="hamburger-container" />
+    <logo class="logo-container" v-if="ctx.device==='mobile'" />
+    <hamburger class="hamburger-container" v-if="ctx.device!=='mobile'" :is-active="ctx.sidebar.opened" @toggleClick="toggleSidebar" />
     <breadcrumb class="breadcrumb-container" v-if="ctx.device!=='mobile'" />
     <div class="right-menu">
-      <el-tooltip content="Under development" effect="dark" placement="bottom">
+      <el-tooltip content="Under development" effect="dark" placement="bottom" v-if="ctx.device!=='mobile'">
         <div class="right-menu-item hover-effect">
           <el-icon>
             <i-ep-Search />
           </el-icon>
         </div>
       </el-tooltip>
-      <div class="right-menu-item hover-effect" @click="drawer = true">
+      <div class="right-menu-item hover-effect" @click="drawerMessage = true" v-if="ctx.device!=='mobile'">
         <el-icon>
           <i-ep-Bell />
         </el-icon>
       </div>
-      <UserCenter />
+      <div class="right-menu-item hover-effect" @click="drawerMenu = true" v-if="ctx.device==='mobile'">
+        <svg-icon icon="menu" />
+      </div>
+      <UserCenter v-if="ctx.device!=='mobile'" />
     </div>
   </div>
 
-  <Teleport to="body">
-    <el-drawer v-model="drawer" title="I am the title" :with-header="true">
+  <Teleport defer to="body">
+    <!-- drawer-menu-container元素是为了:deep样式生效 -->
+    <div class="drawer-menu-container">
+      <el-drawer v-model="drawerMenu" size="min(400px, max(30%, 240px))" :show-close="false">
+        <template #header>
+          <div class="menu-header">
+            <UserCenter size="40px" />
+          </div>
+        </template>
+        <menu-list />
+      </el-drawer>
+    </div>
+  </Teleport>
+
+  <Teleport defer to="body">
+    <el-drawer v-model="drawerMessage" title="I am the title" :with-header="true" size="min(400px, max(30%, 240px))">
       <!-- 历史记录、文档类 -->
       <div style="color:#666;">
         Under development
@@ -33,6 +51,8 @@
 <script setup>
 import { inject, ref } from 'vue'
 
+import MenuList from '../common/Menu/index.vue'
+import Logo from '../common/Logo.vue'
 import Breadcrumb from './Breadcrumb.vue'
 import Hamburger from './Hamburger.vue'
 import UserCenter from '../common/UserCenter.vue'
@@ -40,7 +60,8 @@ import UserCenter from '../common/UserCenter.vue'
 import { dispatch } from '@/store'
 
 const ctx = inject('context')
-const drawer = ref(false)
+const drawerMenu = ref(false)
+const drawerMessage = ref(false)
 
 const toggleSidebar = () => {
   dispatch.sidebar.toggle()
@@ -49,13 +70,17 @@ const toggleSidebar = () => {
 </script>
 
 <style lang="scss" scoped>
+@use '@/styles/variables.module.scss' as v;
 .navbar {
   height: 50px;
   overflow: hidden;
   position: relative;
   background: #fff;
   box-shadow: 0 1px 4px rgba(0, 21, 41, 0.08);
-
+  .logo-container{
+    float: left;
+    margin-left: 15px;
+  }
   .hamburger-container {
     line-height: 46px;
     height: 100%;
@@ -110,4 +135,13 @@ const toggleSidebar = () => {
     }
   }
 }
+.menu-header{
+  height:80px;
+  display: flex;
+  align-items: center;
+}
+
+.drawer-menu-container :deep(.el-drawer__header) {margin-bottom:0;padding-top:0;background:rgba(0,0,0,.05);border-bottom:solid 1px rgba(0,0,0,.05)}
+.drawer-menu-container :deep(.el-drawer__body) {padding: 0 !important}
+.drawer-menu-container :deep(.el-drawer) {background-color: v.$menuBg;}
 </style>