+++ src/layout/components/common/Logo.vue
@@ -1,13 +1,12 @@
 <template>
-  <div :class="{'collapse': props.collapse}" class="sidebar-logo-container">
+  <div class="logo-container" :theme="theme">
     <transition name="sidebarLogoFade">
-      <router-link to="/" v-if="props.collapse">
-        <img class="sidebar-logo" src="@/assets/logo.svg" v-if="showLogo" />
-        <h1 class="sidebar-title" v-else>{{ title }}</h1>
+      <router-link to="/" v-if="collapse">
+        <img class="logo" src="@/assets/logo.svg" />
       </router-link>
       <router-link to="/" v-else>
-        <img class="sidebar-logo" src="@/assets/logo.svg" v-if="showLogo" />
-        <h1 class="sidebar-title">{{ title }}</h1>
+        <img class="logo" src="@/assets/logo.svg" />
+        <h1 class="title">{{ title }}</h1>
       </router-link>
     </transition>
   </div>
@@ -18,54 +17,58 @@ import { ref } from 'vue'
 const props = defineProps({
     collapse: {
         type: Boolean,
-        required: true,
+        default: false,
+    },
+    theme: {
+        type: String,
+        default: 'dark', // dark light
     },
 })
 const title = ref('Vue Admin')
-const showLogo = ref(true)
 </script>
 
 <style lang="scss" scoped>
 .sidebarLogoFade-enter-active {
-    transition: opacity 1.5s;
+  transition: opacity 1.5s;
 }
 
-.sidebarLogoFade-enter-from,
-.sidebarLogoFade-leave-to {
-    opacity: 0;
+// .sidebarLogoFade-leave-to 
+.sidebarLogoFade-enter-from{
+  opacity: 0;
 }
 
-.sidebar-logo-container {
-    position: relative;
-    width: 100%;
-    height: 50px;
+.logo-container {
+  position: relative;
+  // width: 100%;
+  height: 50px;
+  line-height: 50px;
+  text-align: center;
+  overflow: hidden;
+
+  &[theme="dark"] {
+    color: var(--el-color-black);
+  }
+  &[theme="light"] {
+    color: #fff;
+  }
+  
+  .logo {
+    width: 32px;
+    height: 32px;
+    vertical-align: middle;
+    
+  }
+
+  .title {
+    display: inline-block;
+    margin: 0;
+    padding-left: 12px;
+    font-weight: 600;
     line-height: 50px;
-    background: #2b2f3a;
-    text-align: center;
-    overflow: hidden;
-
-    .sidebar-logo {
-        width: 32px;
-        height: 32px;
-        vertical-align: middle;
-        margin-right: 12px;
-    }
-
-    .sidebar-title {
-        display: inline-block;
-        margin: 0;
-        color: #fff;
-        font-weight: 600;
-        line-height: 50px;
-        font-size: 14px;
-        font-family: Avenir, Helvetica Neue, Arial, Helvetica, sans-serif;
-        vertical-align: middle;
-    }
-
-    &.collapse {
-        .sidebar-logo {
-            margin-right: 0px;
-        }
-    }
+    font-size: 14px;
+    font-family: Avenir, Helvetica Neue, Arial, Helvetica, sans-serif;
+    vertical-align: middle;
+  }
+  
 }
 </style>